Move WebcilReader reflection to a helper; add lazy initialization
